requirements.txt
pandas
requests
fpdf2
